Re-work how vars are defined and interacted with
On second look, the 'design-tokens' stuff didn't sit comfortably
because the config as a whole is token-based. To clear things up
I've set it that 'tokens' is now 'css-vars' in the config and to use
them, a collection sets 'css-vars' to true, then changes the 'items'
to be a reference to the 'css-vars' config. Works nice!